requirements.txt
aiohttp-cors==0.7.0 ; python_version >= "3.9" and python_version < "4.0"
aiohttp-jinja2==1.5.1 ; python_version >= "3.9" and python_version < "4.0"
aiohttp==3.8.4 ; python_version >= "3.9" and python_version < "4.0"
aioice==0.7.6 ; python_version >= "3.9" and python_version < "4.0"
aiortc==1.4.0 ; python_version >= "3.9" and python_version < "4.0"
aiosignal==1.3.1 ; python_version >= "3.9" and python_version < "4.0"
anyio==3.6.2 ; python_version >= "3.9" and python_version < "4.0"
async-timeout==4.0.2 ; python_version >= "3.9" and python_version < "4.0"
attrs==22.2.0 ; python_version >= "3.9" and python_version < "4.0"
av==10.0.0 ; python_version >= "3.9" and python_version < "4.0"
beautifulsoup4==4.11.2 ; python_version >= "3.9" and python_version < "4.0"
certifi==2022.12.7 ; python_version >= "3.9" and python_version < "4.0"
cffi==1.15.1 ; python_version >= "3.9" and python_version < "4.0"
charset-normalizer==3.0.1 ; python_version >= "3.9" and python_version < "4.0"
click==8.1.3 ; python_version >= "3.9" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.9" and python_version < "4.0" and sys_platform == "win32" or python_version >= "3.9" and python_version < "4.0" and platform_system == "Windows"
coloredlogs==15.0.1 ; python_version >= "3.9" and python_version < "4.0"
coverage[toml]==7.1.0 ; python_version >= "3.9" and python_version < "4.0"
cryptography==39.0.1 ; python_version >= "3.9" and python_version < "4.0"
dnspython==2.3.0 ; python_version >= "3.9" and python_version < "4.0"
exceptiongroup==1.1.0 ; python_version >= "3.9" and python_version < "3.11"
fastapi==0.92.0 ; python_version >= "3.9" and python_version < "4.0"
filelock==3.9.0 ; python_version >= "3.9" and python_version < "4.0"
flatbuffers==23.1.21 ; python_version >= "3.9" and python_version < "4.0"
frozenlist==1.3.3 ; python_version >= "3.9" and python_version < "4.0"
future==0.18.3 ; python_version >= "3.9" and python_version < "4.0"
gdown==4.6.3 ; python_version >= "3.9" and python_version < "4.0"
google-crc32c==1.5.0 ; python_version >= "3.9" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.9" and python_version < "4.0"
httpcore==0.16.3 ; python_version >= "3.9" and python_version < "4.0"
httptools==0.5.0 ; python_version >= "3.9" and python_version < "4.0"
httpx==0.23.3 ; python_version >= "3.9" and python_version < "4.0"
humanfriendly==10.0 ; python_version >= "3.9" and python_version < "4.0"
idna==3.4 ; python_version >= "3.9" and python_version < "4.0"
iniconfig==2.0.0 ; python_version >= "3.9" and python_version < "4.0"
itsdangerous==2.1.2 ; python_version >= "3.9" and python_version < "4.0"
jinja2==3.1.2 ; python_version >= "3.9" and python_version < "4.0"
markupsafe==2.1.2 ; python_version >= "3.9" and python_version < "4.0"
mpmath==1.2.1 ; python_version >= "3.9" and python_version < "4.0"
multidict==6.0.4 ; python_version >= "3.9" and python_version < "4.0"
netifaces==0.11.0 ; python_version >= "3.9" and python_version < "4.0"
numpy==1.24.2 ; python_version < "4.0" and python_version >= "3.9"
onnxruntime-gpu==1.14.0 ; python_version >= "3.9" and python_version < "4.0"
opencv-python==4.7.0.68 ; python_version >= "3.9" and python_version < "4.0"
packaging==23.0 ; python_version >= "3.9" and python_version < "4.0"
pluggy==1.0.0 ; python_version >= "3.9" and python_version < "4.0"
protobuf==4.22.0 ; python_version >= "3.9" and python_version < "4.0"
pycparser==2.21 ; python_version >= "3.9" and python_version < "4.0"
pydantic==1.10.5 ; python_version >= "3.9" and python_version < "4.0"
pyee==9.0.4 ; python_version >= "3.9" and python_version < "4.0"
pylibsrtp==0.8.0 ; python_version >= "3.9" and python_version < "4.0"
pyopenssl==23.0.0 ; python_version >= "3.9" and python_version < "4.0"
pyreadline3==3.4.1 ; sys_platform == "win32" and python_version >= "3.9" and python_version < "4.0"
pysocks==1.7.1 ; python_version >= "3.9" and python_version < "4"
pytest-aiohttp==1.0.4 ; python_version >= "3.9" and python_version < "4.0"
pytest-asyncio==0.20.3 ; python_version >= "3.9" and python_version < "4.0"
pytest-cov==4.0.0 ; python_version >= "3.9" and python_version < "4.0"
pytest-mock==3.10.0 ; python_version >= "3.9" and python_version < "4.0"
pytest==7.2.1 ; python_version >= "3.9" and python_version < "4.0"
python-dotenv==0.21.1 ; python_version >= "3.9" and python_version < "4.0"
python-multipart==0.0.5 ; python_version >= "3.9" and python_version < "4.0"
pyyaml==6.0 ; python_version >= "3.9" and python_version < "4.0"
requests[socks]==2.28.2 ; python_version >= "3.9" and python_version < "4"
rfc3986[idna2008]==1.5.0 ; python_version >= "3.9" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.9" and python_version < "4.0"
sniffio==1.3.0 ; python_version >= "3.9" and python_version < "4.0"
soupsieve==2.4 ; python_version >= "3.9" and python_version < "4.0"
starlette==0.25.0 ; python_version >= "3.9" and python_version < "4.0"
sympy==1.11.1 ; python_version >= "3.9" and python_version < "4.0"
tomli==2.0.1 ; python_version >= "3.9" and python_full_version <= "3.11.0a6"
tqdm==4.64.1 ; python_version >= "3.9" and python_version < "4.0"
typing-extensions==4.5.0 ; python_version >= "3.9" and python_version < "4.0"
urllib3==1.26.14 ; python_version >= "3.9" and python_version < "4"
uvicorn[standard]==0.20.0 ; python_version >= "3.9" and python_version < "4.0"
uvloop==0.17.0 ; sys_platform != "win32" and sys_platform != "cygwin" and platform_python_implementation != "PyPy" and python_version >= "3.9" and python_version < "4.0"
watchfiles==0.18.1 ; python_version >= "3.9" and python_version < "4.0"
websockets==10.4 ; python_version >= "3.9" and python_version < "4.0"
yarl==1.8.2 ; python_version >= "3.9" and python_version < "4.0"
